requirements.txt
# Core Flask Framework
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-Marshmallow==1.2.1
marshmallow-sqlalchemy==0.29.0
flask-cors==6.0.2

# Database
psycopg[binary]==3.3.3

# Authentication & Security
passlib[bcrypt]==1.7.4
PyJWT==2.11.0

# Configuration
python-dotenv==1.0.1

# Server
gunicorn==21.2.0

# Testing
pytest==9.0.2

# Phase 4: Payments
stripe==14.3.0

# Phase 5: AI and Background Workers
celery==5.6.2
redis==7.2.0
openai==2.21.0
anthropic==0.83.0

# Additional utilities
requests==2.32.5
python-dateutil==2.9.0.post0

# Dependencies automatically installed by Flask ecosystem
# (included for completeness)
alembic==1.18.4
bcrypt==5.0.0
blinker==1.9.0
click==8.3.1
greenlet==3.3.2
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
pluggy==1.6.0
Pygments==2.19.2
six==1.17.0
SQLAlchemy==2.0.46
typing_extensions==4.15.0
Werkzeug==3.1.6